amazon/requirements.txt
Scrapy==1.4.0
PyMySQL==0.7.11
PyDispatcher==2.0.5
